+++ src/models/controlpanel/controlpanel.js
@@ -3,6 +3,7 @@
  * @module models/controlpanel/controlpanel
  */
 
+import { compact, keys, map } from 'lodash';
 import { getRootUrl, translateSchema } from '../../helpers';
 import { Model } from '../../models';
 
@@ -31,5 +32,22 @@ export class Controlpanel extends Model {
     return extend
       ? { ...json, data: this.data, schema: translateSchema(this.schema, req) }
       : json;
+  }
+
+  /**
+   * Get factory fields.
+   * @method getFactoryFields
+   * @static
+   * @param {string} factory Factory field.
+   * @returns {Array} Array of fields with given factory.
+   */
+  getFactoryFields(factory) {
+    const properties = this.schema.properties;
+
+    // Get factory fields
+    const factoryFields = map(keys(properties), (property) =>
+      properties[property].factory === factory ? property : false,
+    );
+    return compact(factoryFields);
   }
 }