+++ src/routes/site/site.js
@@ -3,8 +3,10 @@
  * @module routes/site/site
  */
 
+import { last } from 'lodash';
 import { Controlpanel } from '../../models';
 import { getRootUrl } from '../../helpers';
+const { config } = require(`${process.cwd()}/config`);
 
 export default [
   {
@@ -13,15 +15,17 @@ export default [
     permission: 'View',
     handler: async (req, trx) => {
       const controlpanel = await Controlpanel.fetchById('site', {}, trx);
-      const config = controlpanel.data;
+      const site = controlpanel.data;
 
       // Return database information
       return {
         json: {
           '@id': `${getRootUrl(req)}/@site`,
-          robots_txt: config?.robots_txt,
-          site_logo: config?.site_logo,
-          site_title: config?.site_title,
+          'plone.robots_txt': site?.robots_txt,
+          'plone.site_logo': site?.site_logo
+            ? `${config.frontendUrl}/en/@@images/${site.site_logo.uuid}.${last(site.site_logo.filename.split('.'))}`
+            : null,
+          'plone.site_title': site?.site_title,
         },
       };
     },